+++ src/Symfony/Component/Finder/Finder.php
@@ -48,6 +48,7 @@ class Finder implements \IteratorAggregate, \Countable
     private $depths = array();
     private $sizes = array();
     private $followLinks = false;
+    private $reverseSorting = false;
     private $sort = false;
     private $ignore = 0;
     private $dirs = array();
@@ -461,6 +462,18 @@ class Finder implements \IteratorAggregate, \Countable
     }
 
     /**
+     * Reverses the sorting.
+     *
+     * @return $this
+     */
+    public function reverseSorting()
+    {
+        $this->reverseSorting = true;
+
+        return $this;
+    }
+
+    /**
      * Sorts files and directories by the last inode changed time.
      *
      * This is the time that the inode information was last modified (permissions, owner, group or other metadata).
@@ -736,6 +749,11 @@ class Finder implements \IteratorAggregate, \Countable
 
         if ($this->sort) {
             $iteratorAggregate = new Iterator\SortableIterator($iterator, $this->sort);
+            $iterator = $iteratorAggregate->getIterator();
+        }
+
+        if ($this->reverseSorting) {
+            $iteratorAggregate = new Iterator\ReverseSortingIterator($iterator);
             $iterator = $iteratorAggregate->getIterator();
         }